requirements.txt
click==8.1.3
jsonschema==4.17.3
opentelemetry_exporter_prometheus==1.12.0rc1
opentelemetry-exporter-prometheus-remote-write==0.39b0
prometheus_client==0.17.0
schema==0.7.5
setuptools==67.6.1
sympy==1.12
